feat: support OTEL traces end-to-end
