+++ crates/host/Cargo.toml
@@ -45,7 +45,9 @@ uuid = { workspace = true, features = ["serde"] }
 wascap = { workspace = true }
 wasmcloud-compat = { workspace = true }
 wasmcloud-control-interface = { workspace = true }
+wasmcloud-core = { workspace = true }
 wasmcloud-runtime = { workspace = true }
+wasmcloud-tracing = { workspace = true }
 
 [dev-dependencies]
 toml = { workspace = true, features = ["parse"] }
