+++ crates/tracing/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "wasmcloud-tracing"
+version = "0.1.0"
+
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[features]
+default = []
+otel = ["opentelemetry", "tracing-opentelemetry", "opentelemetry-otlp"]
+
+[dependencies]
+lazy_static = { workspace = true }
+opentelemetry = { workspace = true, features = ["rt-tokio"], optional = true }
+opentelemetry-otlp = { workspace = true, features = [
+    "http-proto",
+    "reqwest-client",
+], optional = true }
+serde = { workspace = true, features = ["derive"] }
+tracing = { workspace = true, features = ["log"] }
+tracing-futures = { workspace = true, features = ["default"] }
+tracing-opentelemetry = { workspace = true, optional = true }
+tracing-subscriber = { workspace = true, features = [
+    "ansi",
+    "env-filter",
+    "fmt",
+    "json",
+] }
+wasmcloud-core = { workspace = true, features = ["otel"] }
